Refactor lrama/lexer.rb
Separate Token class from lrama/lexer.rb
